refactor(panoramique): rename vue component files to fit the naming convention
