+++ plugins/panoramique/tests/component/context-menu.vue
