+++ plugins/panoramique/tests/component/context-option.vue
