+++ plugins/panoramique/tests/component/test-component-composition.vue
