+++ plugins/panoramique/tests/component/test-component-option.vue
