+++ plugins/panoramique/tests/component/test-component.vue
